:recycle: Move nav link components into /layout
