+++ client/src/components/layout/StandardLinks.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
-import styles from "../styles/Navbar.module.css";
-import logo from "../assets/logo.png";
+import styles from "../../styles/Navbar.module.css";
+import logo from "../../assets/logo.png";
 
 const StandardLinks = () => (
   <>
